Create macOS Tor Broswer
Cloudflare has on Mozilla Firefox similar Tor Browser the Cloudflare Purge Plugin. I think I can purge the website out of my traffic in the Cloudflare Account.
What I miss even spoken to Cloudflare is an Account Connection to the webpages which are supported by Cloudflare Captchas. Because there is a problem with the requests in send back. Cloudflare Zero can not manage the process and my website at Cloudflare can not, too. The workers of Cloudflare belong all to my own website. But the intermediale traffic by IP-addresses and browsers give to that no answer.